ImageTool.Worker.AuraSR/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
aura-sr
